requirements.txt
altair==6.0.0
asttokens==3.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.5
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
choreographer==1.2.1
click==8.3.1
colorama==0.4.6
comm==0.2.3
cryptography==46.0.4
curl_cffi==0.13.0
debugpy==1.8.19
decorator==5.2.1
dotenv==0.9.9
executing==2.2.1
frozendict==2.4.7
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
google-auth-oauthlib==1.2.4
gspread==6.2.1
idna==3.11
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kaleido==1.2.0
logistro==2.0.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
multitasking==0.0.12
narwhals==2.15.0
nest-asyncio==1.6.0
numpy==2.4.1
oauthlib==3.3.1
orjson==3.11.6
packaging==26.0
pandas==2.3.3
parso==0.8.5
pathlib==1.0.1
patsy==1.0.2
peewee==3.19.0
pillow==12.1.0
platformdirs==4.5.1
plotly==6.5.2
plotly-express==0.4.1
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==6.33.4
psutil==7.2.1
pure_eval==0.2.3
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydeck==0.9.1
Pygments==2.19.2
pytest==9.0.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
rsa==4.9.1
scipy==1.17.0
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
soupsieve==2.8.3
stack-data==0.6.3
statsmodels==0.14.6
streamlit==1.53.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.4.0
websockets==16.0
yfinance==1.1.0
